fix: исправил ререндер на клиенте при переключении dropdown
